pyproject.toml
[project]
name = "automation-khs"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "elskow", email = "helmyl.work@gmail.com"},
]
dependencies = [
    "selenium>=4.16.0",
    "python-dotenv>=1.0.0",
    "webdriver-manager>=4.0.1",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.dev-dependencies]
format = ["black" , "isort"]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.scripts]
dev = "src/scrapper.py"
format = {shell = "black . && isort ."}

[tool.pdm]
package-type = "automation"
